Reorganize everything into sub-projects
