+++ src/main/scala/boot/CustomBootPin.scala
@@ -1,4 +1,4 @@
-package testchipip
+package testchipip.boot
 
 import chisel3._
 import chisel3.util._
@@ -9,9 +9,6 @@ import freechips.rocketchip.tilelink._
 import freechips.rocketchip.devices.tilelink._
 import freechips.rocketchip.regmapper._
 import freechips.rocketchip.subsystem._
-import freechips.rocketchip.util._
-import freechips.rocketchip.tile._
-import freechips.rocketchip.prci._
 
 case class CustomBootPinParams(
   customBootAddress: BigInt = 0x80000000L, // Default is DRAM_BASE
