+++ src/main/scala/clocking/ClockUtil.scala
@@ -1,4 +1,4 @@
-package testchipip
+package testchipip.clocking
 
 import chisel3._
 import chisel3.util._
