+++ src/main/scala/cosim/Cospike.scala
@@ -1,4 +1,4 @@
-package testchipip
+package testchipip.cosim
 
 import chisel3._
 import chisel3.experimental.{IntParam, StringParam, IO}
