+++ src/main/scala/dram/SimDRAM.scala
@@ -1,4 +1,4 @@
-package testchipip
+package testchipip.dram
 
 import chisel3._
 import chisel3.experimental.IntParam
