+++ src/main/scala/serdes/Configs.scala
@@ -1,4 +1,4 @@
-package testchipip
+package testchipip.serdes
 
 import chisel3._
 import org.chipsalliance.cde.config.{Parameters, Config}
