+++ src/main/scala/serdes/Serdes.scala
@@ -1,4 +1,4 @@
-package testchipip
+package testchipip.serdes
 
 import chisel3._
 import chisel3.util._
