+++ src/main/scala/soc/NoDebug.scala
@@ -1,4 +1,4 @@
-package testchipip
+package testchipip.soc
 
 import chisel3._
 import freechips.rocketchip.subsystem.BaseSubsystem
