+++ src/main/scala/soc/Scratchpad.scala
@@ -1,4 +1,4 @@
-package testchipip
+package testchipip.soc
 
 import chisel3._
