+++ src/main/scala/spi/SPIFlash.scala
@@ -1,4 +1,4 @@
-package testchipip
+package testchipip.spi
 
 import chisel3._
 import chisel3.util.{HasBlackBoxResource}
