+++ src/main/scala/tsi/SimTSI.scala
@@ -1,4 +1,4 @@
-package testchipip
+package testchipip.tsi
 
 import chisel3._
 import chisel3.util._
