+++ src/main/scala/tsi/tsiHost/TsiHostPeriphery.scala
@@ -1,4 +1,4 @@
-package testchipip
+package testchipip.tsi
 
 import chisel3._
 import chisel3.util._
