+++ src/main/scala/util/HeaderEnum.scala
@@ -1,4 +1,4 @@
-package testchipip
+package testchipip.util
 
 import chisel3._
 import chisel3.util.log2Up
