+++ src/main/scala/util/Switcher.scala
@@ -1,4 +1,4 @@
-package testchipip
+package testchipip.util
 
 import chisel3._
 import chisel3.util._
